add ui and sign up
