+++ app/index.tsx
@@ -1,80 +1,39 @@
 import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 import { useRouter } from 'expo-router';
-import { StyleSheet, TextInput, Text, TouchableOpacity, View, Image } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity, View, Image } from 'react-native';
 import React from 'react';
-import api from '../API/api';
-import AsyncStorage from '@react-native-async-storage/async-storage';
 
-/* 
-NEW: index main entry screen
-*/
-
-export default function LoginPage() {
+export default function IndexPage() {
   const router = useRouter();
-  const [username, onChangeUser] = React.useState('');
-  const [password, onChangePassword] = React.useState('');
-  const [errorMessage, setErrorMessage] = React.useState('');
-
-  async function handleSubmit() {
-    console.log('Submitting:', username, password);
-    try {
-      const response = await api.post('/users/authenticate', { username, password });
-      const userId = response.data.userId;
-      await AsyncStorage.setItem('userId', userId.toString()); //save userId in AsyncStorage
-      setErrorMessage('');
-      router.push('/main'); //goes to main
-    } catch (error) {
-      setErrorMessage('Invalid Credentials! Try Again.');
-    }
-  }
 
   return (
     <SafeAreaProvider>
       <SafeAreaView style={styles.container}>
+        {/* Logo */}
         <Image
-          source={require('../assets/placenet.png')} 
+          source={require('../assets/placenet.png')}
           style={styles.logo}
         />
+
+        {/* App Title */}
         <Text style={styles.titleText}>Welcome to Placenet!</Text>
         <Text style={styles.subtitleText}>For property and community care</Text>
 
-        {/* Username Input */}
-        <TextInput
-          style={styles.input}
-          onChangeText={onChangeUser}
-          value={username}
-          placeholder="Username"
-          placeholderTextColor="#A9A9A9"
-        />
-
-        {/* Password Input */}
-        <TextInput
-          style={styles.input}
-          onChangeText={onChangePassword}
-          value={password}
-          placeholder="Password"
-          placeholderTextColor="#A9A9A9"
-          secureTextEntry
-        />
-
-        {/* Error Message */}
-        {errorMessage ? <Text style={styles.errorText}>{errorMessage}</Text> : null}
-
-        {/* Log In Button */}
-        <TouchableOpacity style={styles.button} onPress={handleSubmit}>
-          <Text style={styles.buttonText}>Log In</Text>
+        {/* Sign In Button */}
+        <TouchableOpacity
+          style={styles.button}
+          onPress={() => router.push('/sign_in')}
+        >
+          <Text style={styles.buttonText}>Sign In</Text>
         </TouchableOpacity>
 
-        {/* Sign Up Link */}
-        <View style={styles.signupContainer}>
-          <Text style={styles.signupText}>Don't have an account?</Text>
-          <Text
-            style={styles.signupLink}
-            onPress={() => router.push('/sign_up')} // TODO: add this page 
-          >
-            Sign Up
-          </Text>
-        </View>
+        {/* Sign Up Button */}
+        <TouchableOpacity
+          style={styles.buttonOutline}
+          onPress={() => router.push('/sign_up')}
+        >
+          <Text style={styles.buttonOutlineText}>Sign Up</Text>
+        </TouchableOpacity>
       </SafeAreaView>
     </SafeAreaProvider>
   );
@@ -82,10 +41,10 @@ export default function LoginPage() {
 
 const styles = StyleSheet.create({
   container: {
-    backgroundColor:'#ffffff',
+    backgroundColor: '#ffffff',
     flex: 1,
-    justifyContent:'center',
-    alignItems:'center',
+    justifyContent: 'center',
+    alignItems: 'center',
     padding: 15,
   },
   logo: {
@@ -94,31 +53,17 @@ const styles = StyleSheet.create({
     marginBottom: 20,
   },
   titleText: {
-    fontSize: 24,
+    fontSize: 28,
     fontWeight: 'bold',
     color: '#000',
     marginBottom: 5,
+    textAlign: 'center',
   },
   subtitleText: {
     fontSize: 16,
     color: '#666',
-    marginBottom: 30,
-  },
-  input: {
-    backgroundColor: '#F0F0F0',
-    borderColor: '#C0C0C0',
-    borderRadius: 5,
-    height: 50,
-    width: '80%',
-    margin: 12,
-    paddingHorizontal: 10,
-    fontSize: 16,
-    color: '#000',
-  },
-  errorText: {
-    color: 'red',
+    marginBottom: 40,
     textAlign: 'center',
-    marginBottom: 10,
   },
   button: {
     backgroundColor: '#404040ff',
@@ -127,27 +72,26 @@ const styles = StyleSheet.create({
     width: '80%',
     justifyContent: 'center',
     alignItems: 'center',
-    marginVertical: 20,
+    marginVertical: 10,
   },
   buttonText: {
     color: '#FFFFFF',
     fontSize: 18,
     textAlign: 'center',
   },
-  signupContainer: {
-    flexDirection: 'row',
+  buttonOutline: {
+    borderColor: '#404040ff',
+    borderWidth: 1,
+    borderRadius: 5,
+    height: 50,
+    width: '80%',
     justifyContent: 'center',
     alignItems: 'center',
-    marginTop: 10,
+    marginVertical: 10,
   },
-  signupText: {
-    color: '#A9A9A9',
-    fontSize: 14,
-  },
-  signupLink: {
-    color: '#0000FF',
-    fontSize: 14,
-    marginLeft: 5,
-    textDecorationLine: 'underline',
+  buttonOutlineText: {
+    color: '#404040ff',
+    fontSize: 18,
+    textAlign: 'center',
   },
 });
