+++ package.json
@@ -16,38 +16,40 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
-    "@react-native-async-storage/async-storage": "^2.0.0",
+    "@react-native-async-storage/async-storage": "1.23.1",
     "@react-native-community/datetimepicker": "^8.2.0",
     "@react-navigation/native": "^6.0.2",
     "axios": "^1.7.7",
     "cocoapods": "^0.0.0",
-    "expo": "^51.0.38",
-    "expo-constants": "~16.0.2",
-    "expo-document-picker": "~12.0.2",
-    "expo-font": "~12.0.9",
-    "expo-linking": "~6.3.1",
-    "expo-router": "^3.1.0",
-    "expo-splash-screen": "~0.27.5",
-    "expo-status-bar": "~1.12.1",
-    "expo-system-ui": "~3.0.7",
-    "expo-web-browser": "~13.0.3",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.74.5",
-    "react-native-gesture-handler": "~2.16.1",
-    "react-native-reanimated": "~3.10.1",
-    "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "3.31.1",
-    "react-native-web": "~0.19.10"
+    "expo": "^52.0.7",
+    "expo-constants": "~17.0.3",
+    "expo-document-picker": "~13.0.1",
+    "expo-font": "~13.0.1",
+    "expo-linking": "~7.0.2",
+    "expo-router": "~4.0.5",
+    "expo-splash-screen": "~0.29.10",
+    "expo-status-bar": "~2.0.0",
+    "expo-system-ui": "~4.0.3",
+    "expo-web-browser": "~14.0.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-native": "0.76.2",
+    "react-native-gesture-handler": "~2.20.2",
+    "react-native-reanimated": "~3.16.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.0.0",
+    "react-native-web": "~0.19.10",
+    "expo-secure-store": "~14.0.0",
+    "react-native-svg": "15.8.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@testing-library/react-native": "^12.7.2",
     "@types/jest": "^29.5.12",
-    "@types/react": "~18.2.45",
+    "@types/react": "~18.3.12",
     "@types/react-test-renderer": "^18.0.7",
     "jest": "^29.2.1",
-    "jest-expo": "~51.0.3",
+    "jest-expo": "~52.0.1",
     "react-test-renderer": "18.2.0",
     "typescript": "~5.3.3"
   },
